Move Otel shim injector to its own tooling sub-package
